requirements.txt
PyQt5==5.15.9
pandas==2.1.0
openpyxl==3.1.2
xlrd==2.0.1
pillow==10.0.1
bcrypt==4.0.1
qtawesome==1.2.3